Modified to double precision parameters

+++ Molecular/src/Molecular/Physics/BoundingBox.h
@@ -5,7 +5,7 @@
 namespace Molecular {
     class BoundingBox {
     public:
-        BoundingBox(glm::vec2 minPoint, glm::vec2 maxPoint)
+        BoundingBox(glm::dvec2 minPoint, glm::dvec2 maxPoint)
                 : m_minPoint(minPoint), m_maxPoint(maxPoint) {}
 
         bool IsInside(const glm::vec2 &position) const {
@@ -13,11 +13,11 @@ namespace Molecular {
                    position.y >= m_minPoint.y && position.y <= m_maxPoint.y;
         }
 
-        glm::vec2 GetMinPoint() const { return m_minPoint; }
-        glm::vec2 GetMaxPoint() const { return m_maxPoint; }
+        glm::dvec2 GetMinPoint() const { return m_minPoint; }
+        glm::dvec2 GetMaxPoint() const { return m_maxPoint; }
 
     private:
-        glm::vec2 m_minPoint;  // Lower-left corner of the bounding box
-        glm::vec2 m_maxPoint;  // Upper-right corner of the bounding box
+        glm::dvec2 m_minPoint;  // Lower-left corner of the bounding box
+        glm::dvec2 m_maxPoint;  // Upper-right corner of the bounding box
     };
 }
